Move test to MULANGT11Codop2Test
